requirements.txt
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
six==1.11.0
